requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
cachetools==4.2.2
certifi==2025.1.31
chardet==3.0.4
distro==1.9.0
emoji==2.14.1
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
jiter==0.8.2
nest-asyncio==1.6.0
openai==1.61.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
rfc3986==1.5.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
tornado==6.1
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
